backend/requirements/base.txt
aerich==0.7.1
asyncpg==0.27.0
cryptography==38.0.3
fastapi==0.87.0
httpx==0.23.0
nanoid==2.0.0
protobuf<=3.20.1
pydantic[email]==1.10.2
pyowm==3.3.0
python-jose==3.3.0
python-json-logger==2.0.4
tortoise-orm==0.19.2
uvicorn[standard]==0.19.0
